requirements.txt
PyQt5~=5.15.0
peewee~=3.13.3
mysql-connector-python~=8.0.20
